feat: add lyrics rendering to lockscreen and improve power menu UI
- Introduced a new `lyrics` module for rendering song lyrics on the lockscreen.
- Updated `power_menu` component to enhance visual feedback with improved hover and active states.
- Refactored `lockscreen` to integrate lyrics display and adjust layout for better aesthetics.
- Enhanced clipboard service to support image previews and added functionality for image handling.
- Updated volume control to use floating-point values for more precise volume representation.
- Added a new password input component for Polkit authentication with improved styling and error handling.
- Updated language files to include new strings for image items in clipboard history.
- Improved system service to notify on audio changes and refresh audio status.

+++ crates/app/src/capsule/widgets/volume/volume_bar.rs
@@ -3,7 +3,7 @@ use ui::components::volume_control::VolumeControlBar;
 use ui::theme::Theme;
 
 pub fn render_volume_bar(
-    volume: u32,
+    volume: f32,
     is_muted: bool,
     muted_label: impl Into<SharedString>,
     _theme: &Theme,
